--- BE/package.json
@@ -1,18 +1,0 @@
-{
-  "name": "be",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node src",
-    "dev": "nodemon --watch src/ src/index.js"
-  },
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "@koa/cors": "^3.1.0",
-    "koa": "^2.13.0",
-    "nodemon": "^2.0.4"
-  }
-}
