--- BE/src/api/area/index.js
@@ -1,51 +1,0 @@
-const Router = require('koa-router');
-const area =  new Router();
-
-const dummy = [
-  { code: "cv", score: 90, cnt: 55 },
-  { code: "cf", score: 40, cnt: 12 },
-  { code: "rs", score: 74, cnt: 40 },
-  { code: "hp", score: 10, cnt: 4 },
-  { code: "sm", score: 60, cnt: 2 },
-  { code: "pc", score: 90, cnt: 24 },
-]
-
-/* address로 세권 찾기
-  GET /api/area/address
-  { address, selectedList }
-*/
-area.get('/address', (ctx) => {
-  const address = ctx.query.address;
-  const selectedList = JSON.parse(ctx.query.selectedList)
-  
-  if (address && selectedList.length) {
-    ctx.status = 200;
-    ctx.body = { 
-      address,
-      data: dummy.filter(x => selectedList.includes(x.code))
-    }
-  } else {
-    ctx.status = 400
-  }
-});
-
-/* GPS정보로 세권 찾기
-  GET /api/area/geo
-  { geo, selectedList }
-*/
-area.get('/geo', (ctx) => {
-  const geoloc = ctx.query.address;
-  const selectedList = JSON.parse(ctx.query.selectedList)
-
-  if (geoloc && selectedList.length) {
-    ctx.status = 200;
-    ctx.body = { 
-      geoloc,
-      data: dummy.filter(x => selectedList.includes(x.code))
-    }
-  } else {
-    ctx.status = 400
-  }
-});
-
-module.exports = area;
