+++ src/components/ResultChart.css
