+++ src/components/SearchResult.css
